Change the default 'version_regexp' to catch a wider variety of tag patterns
A small modification to the default allowed it to catch the original style
as well as a large number of additional styles.

It felt dumb to me that specifying an alternate 'match_pattern'
required an alternate 'version_regexp'.
That is no longer strictly necessary.

Also, for additional ease,
instead of requiring the pattern to be anchored at both ends
just do a simple match against the regexp and then set the whole thing to $1.